requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.16
marshmallow==3.14.1
razorpay==1.3.0
setuptools
pymysql
cryptography
flask_migrate
flask-bcrypt
python-dotenv
PyJWT
pytz